Bump minimum CMake version required

+++ synchrono/intersection.cpp
@@ -68,7 +68,7 @@ double step_size = 3e-3;
 double render_step_size = 1.0 / 50;  // FPS = 50
 
 // How often SynChrono state messages are interchanged
-float heartbeat = 1e-2;  // 100[Hz]
+float heartbeat = 1e-2f;  // 100[Hz]
 
 void AddCommandLineOptions(ChCLI& cli);
 void LogCopyright(bool show);
@@ -104,7 +104,7 @@ int main(int argc, char* argv[]) {
     // Normal simulation options
     step_size = cli.GetAsType<double>("step_size");
     end_time = cli.GetAsType<double>("end_time");
-    heartbeat = cli.GetAsType<double>("heartbeat");
+    heartbeat = cli.GetAsType<float>("heartbeat");
 
     const bool use_sensor = cli.HasValueInVector<int>("sens", node_id);
     const bool sensor_vis = cli.GetAsType<bool>("sens_vis");
